Fix hangs on git-upload-pack with shallow checkout
The `git-upload-pack` binary will remain open waiting for further
instructions on `stdin` in some cases (e.g. with shallow checkouts
when you pass `--depth 1`).

This results in the operation hanging indefinitely because the client
has already sent all of its data, but the server keeps waiting. The
stdin pipe is now immediately closed once data is copied from it.

Additionally, the pipe that comes back from `gitCommand` is now also
an `io.ReadCloser` so that we can ensure it doesn't leak. Normally,
it's closed for us by `exec.Cmd::Wait()`, but it's possible to return
early before that's called.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>